radio fixings
