--- src/components/NavPanel/nav-panel-old.styl
@@ -1,12 +1,0 @@
-@import 'variables.styl'
-
-.nav-panel
-  -ms-overflow-style: -ms-autohiding-scrollbar
-  -webkit-overflow-scrolling: touch
-  background-color: $tab-panel-background
-  color: alpha(black 75%)
-  display: block
-  overflow-x: scroll
-  position: relative
-  white-space: nowrap
-  width: 100%
